Fixed tower attack

+++ src/Plugins/Plugin/World/Tower/TowerManager.h
@@ -18,7 +18,7 @@ public:
 
 	void		Update			(float dt);
 
-	void		CreateTower		(TowerBase::ETowerType towerType, TowerBase::EFocusType focusType, const CShVector3 & position, float damages, float attackSpeed);
+	void		CreateTower		(TowerBase::ETowerType towerType, TowerBase::EFocusType focusType, const CShVector3 & position, int damages, float attackSpeed);
 
 	void		GetTowerList	(CShArray<TowerBase *>& aTowerList);
